third_party/blink/web_tests/css3/selectors3/xml/css3-modsel-15c-expected.txt

CONSOLE ERROR: Uncaught TypeError: document.getElementsByTagNameNS(...)[0].setIdAttribute is not a function
